Fix for the new FSize class (avoid the 8EiB limit, bsc#991090)
- Update the package version
- 2.49.5

+++ libyui-qt/src/QY2DiskUsageList.cc
@@ -278,7 +278,7 @@ QY2DiskUsageListItem::usedPercent() const
     int percent = 0;
 
     if ( totalSize() != 0 )
-	percent = ( 100 * usedSize() ) / totalSize();
+	percent = int(( 100 * usedSize() ) / totalSize());
 
     return percent;
 }
